model_builds/triposr/requirements_triposr_server_gpu.txt
--extra-index-url https://download.pytorch.org/whl/cu121

# Core PyTorch stack (CUDA 12.1 toolchain)
torch==2.2.2+cu121
torchvision==0.17.2+cu121
torchaudio==2.2.2+cu121

# Image and geometry utilities
numpy>=1.23.0,<2.0
Pillow>=9.0.0
pybind11>=2.10.0
trimesh>=4.0.0

# TripoSR dependencies
scikit-learn>=1.0.0
scikit-build-core>=0.7.0
transformers>=4.41.0,<5.0
accelerate>=0.21.0
safetensors>=0.4.0
huggingface_hub>=0.20.0,<0.24.0
triposr @ https://github.com/AO-StreetArt/TripoSR/archive/refs/tags/v1.0.0.tar.gz

# HTTP serving
Flask>=2.3.0
gunicorn>=21.2.0
